Fix raven staff in close range
The raven staff's untomed missile has a very small delay before it splits into 3 (and then a longer delay to split into 5). If it's fired in melee range, it explodes before splitting in 3, so 2/3rds of the potential damage is lost. Now it checks if there's a monster in front of us, and if so split into 3 immediately (otherwise use original delay).